Adding some more :ssize_t casts to IO.chpl
The move of miniMD to the release/examples directory caused it to be
tested on 32-bit linux for the first time, and revealed new paths in
the IO module that assumed int(64)s would match ssize_t's.  I took
the cheap approach of downcasting them to try and clean up testing
tonight, and will ask Michael to review after the fact.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21926 3a8e244f-b0f2-452b-bcba-4c88e055c3ca