Adding support for aries conduit on XC
[authored by Greg, tested and reviewed by me]

These are the changes required on XC to enable GASNet Aries conduit
support.  I was investigating Ryan build issues this morning which
were somewhat related, and so used the opportunity to review and test
Greg's patch.  It seems to be working for me (which is to say, a
multi-locale hello world worked).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21930 3a8e244f-b0f2-452b-bcba-4c88e055c3ca