Include LLVM and Clang tarballs in Chapel SVN and releases
[reviewed by Lydia]

Instead of the old strategy of downloading LLVM and Clang with SVN, we include
a .tar.gz for each and unpack them when building LLVM support. This way, the
Chapel svn checkout or release contains everything that is needed, and SVN
is tracking artifacts that we rely upon (as it does with GASNet for example).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21934 3a8e244f-b0f2-452b-bcba-4c88e055c3ca