Replaced checkForMissingDefs() with a more efficient addition to verify().
[reviewed by Tom]

Specifically, an addition to DefExpr::verify(), together with existing
code in SymExpr::verify() cover the checks in checkForMissingDefs().

Incorrect removal of DefExprs in deadBlockRemoval() was causing
one of the assertion to be violated. That was fixed by r21849,
so we can now add this check to DefExpr::verify().

While there - changed the scope of checkNoRecordDeletes() from global to
static within checkResolved.cpp.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21939 3a8e244f-b0f2-452b-bcba-4c88e055c3ca