Use sublocales in DefaultRectangular and range leader iterators
[Reviewed by Sung]

If there are sublocales, do a coforall+on over them and a coforall over their
cores to distribute the iterations for DefaultRectangular and for ranges.

For numa, this breaks the iterations up across the domains and also runs in
parallel on multicore domains.

If the locale model doesn't have sublocales the behavior is unchanged.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21944 3a8e244f-b0f2-452b-bcba-4c88e055c3ca