Retiring test/parallel/begin/sungeun/capture.future.
[requested by Brad]

This feature request (capture for tasks) was implemented in r21940.

As a follow up to r21940, here are some early discussion notes
regarding this feature, from the being-removed .future:

---
Brad suggested that the existence of task private variables could make
the situation better/clearer.

Evaluating some of the begin block at task creation time opens up a
bit of a can of worms because there is a question of where to draw the
line (how much should be evaluated?).

Another option might be to go the style of XMT futures where any
non-global variables that are used in the block are required to be
"passed in".  This might allow us to get rid of the heap allocation
step we have now during the parallel pass.

The solution should probably apply to cobegins also.
---


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21947 3a8e244f-b0f2-452b-bcba-4c88e055c3ca