Moved checkResolveRemovedPrims() into verify
[trivial, not reviewed]

The checkResolveRemovedPrims() call is arguably a verify call
that makes more sense in the new checks.cpp than embedded at
the end of functionResolution.cpp, so moved it there.  My
real reason for doing so, however, was that when using 
--ignore-errors-for-pass and using certain PRIMS, it causes
an internal error to fire, which makes that flag less useful
for testing.  By moving it into verify, tests can be made
to exit more cleanly, even if they use these primitives.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21948 3a8e244f-b0f2-452b-bcba-4c88e055c3ca