Add membership tests via varargs as well as tuples
[contributed by Sara Salem, reviewed and tweaked by me]

Prior to this change, membership tests on domains required
the argument to be a tuple value for a multidimensional
domain or array.   This came about in a roundabout way...

Sara Salem, a student at Cairo University, was asking about the
existence of a membership test on (sparse?) domains.  Checking, we
were surprised that the sparse domain primer (if not others) didn't
mention this feature, so I invited her to add it and submit a patch.
When she did, we realized that the code would be much cleaner if
membership accepted a varargs list of integers, much like a normal
array access.  So she took that change on as well.

This commit does not include one change she proposed which was to add
an error overload for illegal cases.  While this produced nice error
messages in many cases, it had the downside of not allowing membership
tests to be promoted (the promotion would be a better match against
the error case), so I removed it.

This commit includes the new overloads in ChapelArray.chpl, the
changes to the primer, and two tests created by Sara.  One tests a
number of legal cases; the second is a modification of that I created
which tests a number of illegal cases and includes her error overload
within the test itself (to avoid the promotion issue in this one local
place).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21949 3a8e244f-b0f2-452b-bcba-4c88e055c3ca