Feature request for last-resort function resolution
This is a feature request that's come up a few times in the past six
months or so (the most recent being last night's commit of Sara's
member() improvements, which had originally included a draft for a
routine that would benefit from the feature).  It asks for an ability
to mark a given function overload as a "last resort" or "catch-all" or
"error case" to indicate that it should only be considered in the
event that no other function provides a good match.  I motivate this
with a few examples that have come up when working in standard
modules, and document the idea in lastResort.future.

My suspicion is that this would not be a difficult feature to
implement, other than the fact that it requires some surgery in the
thick of function resolution.  Finding the time on the other hand...



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21952 3a8e244f-b0f2-452b-bcba-4c88e055c3ca