Allow variables used in PRIM_SIZEOF and PRIM_CAST_TO_VOID_STAR to be narrowed
[Reviewed by David]

The use of PRIM_CAST_TO_VOID_STAR due to the chpl_here_alloc rework
(r21945) caused a test failure and exposed the fact that we were not
allowing variable used as arguments to PRIM_CAST_TO_VOID_STAR to be
narrowed.  Added PRIM_SIZEOF also, but it's possible there are others.

Note: The reason this test failed is because we failed to narrow a
reference to the extern const chpl_nullTaskID.  In the runtime this is
really a #define, so the this led to a C compilation error when trying
to use chpl_nullTaskID as an lvalue.  In a more normal cases, we
probably wouldn't have noticed.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21978 3a8e244f-b0f2-452b-bcba-4c88e055c3ca