Adding one more binary trees variation
While reviewing the binary trees written specification, I noted that
it appears to be legal to free tree nodes while computing checksums
rather than in a separate pass.  I'm not confident that this is legal,
but it seemed worthwhile to create a version that took this approach
to see how the performance compares.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21982 3a8e244f-b0f2-452b-bcba-4c88e055c3ca