Throwing -Wno-inline for GASNet builds
Michael and Tom were correct and I was mistaken: It is GASNet's
"optimization" flags that are causing -Winline to be thrown.  The last
time I looked into whether we used GASNet's choices of flags, it seems
that I only looked in the runtime which is why I was mistaken: We
don't use GASNet flags for building our runtime, but we do for the
generated code.





git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21985 3a8e244f-b0f2-452b-bcba-4c88e055c3ca