Added TODO to generated code's Makefile.include
While implementing the -Wno-inline issue, I noticed for the first time
that the compiler line we use to build a generated executable has
become pretty ridiculously long and redundant.  From what I can tell,
this is because at some point (I think mppf's 19641) we decided to add
all of the RUNTIME_CFLAGS to the GEN_CFLAGS variable which already had
many of the same shared sub-variables -- so, for example, all of the
warning flags now appear twice.  

I believe the better fix would be to tease the parts of RUNTIME_CFLAGS
that we think we need in both cases (and from the comments, it appears
to be -D flags to define tasking models and such) into its own
Makefile variable and then including this variable in both
RUNTIME_CFLAGS and GEN_CFLAGS rather than just having GEN_CFLAGS
include all of RUNTIME_CFLAGS (since they were architected to be
siblings, along with COMPILER_CFLAGS).

This was a bigger change than I could take on prior to the release,
and arguably just for aesthetics (or ability to make sense of the cc
line used to compile generated code) so I didn't.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21987 3a8e244f-b0f2-452b-bcba-4c88e055c3ca