Made 'aprun' the default launcher for Cray systems
Prior to this commit, we used qsub+aprun as the default launcher for
Cray systems when both were in this path.  However, we've recently
been running into issues with the PBS launcher that have caused more
trouble than it is perhaps worth to put on a user as a default,
particularly when many systems are configured to support interactive
aprun approaches.

We've discussed "smarter" ways for determining the default launcher,
for example checking to see if the pbs is a version we know and/or
checking to see whether the system permits interactive runs.  However,
for now we're going to follow the KISS principle and set it to aprun
by default (though users can still override this default if they want
to use the pbs+aprun launcher).



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21988 3a8e244f-b0f2-452b-bcba-4c88e055c3ca