Backing out r21993
r21993 duplicated the effort of r21987 and was actually the first way
I tried to fix the issue.  It isn't effective because the -Winline
flag comes from GASNet and GASNet's flags are appended after the
WARN_GEN_CFLAGS are; and it isn't necessary here because -Wall doesn't
imply -Winline.  So the r21987 approach is better by only overriding
the flag when necessary and in the correct position relative to the
GASNet flags.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21994 3a8e244f-b0f2-452b-bcba-4c88e055c3ca