Improve the removeUnnecessaryAutoCopyCalls optimization. [reviewed by hilde] [committed on behalf of Alexey Gokhberg]
Using the following simple Chapel program for testing:
 
config const N = 50000;
 
proc main() {
    var A, B, C: [1..N] real;
    (B, C) = (2.0, 1.0);
    const alpha = 3.0;
    forall (a, b, c) in zip(A, B, C) do
        a = b * alpha + c;
}
 
the Chapel compier generates rather long chains of matching chpl__autoCopy / ~chpl_destroy call pairs that the original removeUnnecessaryAutoCopyCalls pass is not able to optimize for the reasons discussed earlier. 
 
The calls in question have the following origin:
 
* chpl__autoCopy calls are produced by instantiate_tuple_autoCopy() defined in generics.c and called at pass buildDefaultFunctions. This function generates chpl__autoCopy for all tuple elemens with no regards to their type. The compiler does not care of generating the matching chpl__autoDestroy at the end of the elements' life.
 
* ~chpl_destroy calls are produced by fixupDestructors() defined in callDestructors.cpp and called at pass callDestructors. This function generates ~chpl_destroy for the destruction of aggegate data types that are not classes and also for sync/single objects.
 
In our case dring the aforementioned passes the data types are represented by array wrapper records. These records are in turn members of 3-element tuples - hence use of instantiate_tuple_autoCopy().
 
The quick fix proposed in my previous mail did not pass the regression test. Therefore more systematic solution is required. If we want the original removeUnnecessaryAutoCopyCalls algorithm to work as designed, we should modify fixupDestructors() in order to produce chpl__autoDestroy instead of ~chpl_destroy for types that have FLAG_REMOVABLE_AUTO_DESTROY set, that is, for warpper records of distributions, domains, and arrays. It is worth noting that for these particular data types ~chpl_destroy  and chpl__autoDestroy have identical functuionality and thus one can be safely replaced with another (see details in ChapelBase.chpl, note also that this statement is not necessarily correct for the arbitrary data types).
 
Therefore the starting section of fixupDestructors() was modified to perform that substitution.
 
Unfortunately, this change alone is not sufficient to force the removeUnnecessaryAutoCopyCalls algorithm to work as intended. To be effective, this algorithm must be used with wrapper records removed, that is, after the completion of pass removeWrapRecords.  To be most effective, removeUnnecessaryAutoCopyCalls has to be moved in the pass list right after removeWrapRecords. 

With this change, results look reasonable. However, the generated C code is more compact if we retain also the original removeUnnecessaryAutoCopyCalls instance of and add copyPropagation and deadCodeElimination after the relocated removeUnnecessaryAutoCopyCalls instance.  However, due to interaction with other passes, this change had to be reverted, so removeUnnecessaryAutoCopyCall appears at its original location in the pass list.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21995 3a8e244f-b0f2-452b-bcba-4c88e055c3ca