Fix auto-detection of compiler for Cray platforms
[trivial, not reviewed, though I'd appreciate a post-commit review
from Michael]

This fixes a bug with the auto-detection of which compiler to
use on cray-x* platforms that crept in during r21803.  

In particular, traditionally we have passed the --host/--target flag
supplied to the compiler query along to query the platform in order to
guess defaults based on platform.  If both CHPL_HOST_PLATFORM and
CHPL_TARGET_PLATFORM are set to "cray-x*" (the recommended mode for
end users) then the bug isn't seen.  However, if CHPL_HOST_PLATFORM is
set to linux64 (for a cross-compilation environment, for example),
then failing to pass along the --host flag causes the script to
determine the target compiler to be 'gnu' rather than
'cray-prgenv-cray' (or whichever PrgEnv is loaded).

This is why whitebox testing for the cray-xt has been broken since
early September, re-running the same test configuration against
vanilla 'gnu' over and over again in spite of different PrgEnv-*
modules being loaded.




git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@21998 3a8e244f-b0f2-452b-bcba-4c88e055c3ca