Fixed a bug in the implementation of "capture for tasks" of r21940. Also clarified capturing for coforall index variables.
THE BUG was that certain variables that should be captured, weren't.

Let's say we are compiling this source program:

  iRef:int comes from outside 'begin'
  ...
  begin {
    iRef = 5;      // (a)
    writeln(iRef); // (b)
  }

With capture-for-tasks, (a) is illegal. In the absence of that, (b)
should use the captured copy of the original iRef.

Before this change it worked -- except if, at resolution time, iRef is a
reference, i.e., _ref(int(64)). This can happen, e.g., if iRef
is a 'ref' formal or an index variable of a loop with a 'var' iterator.
Capturing is still required, but was not happening.

In this case, the begin_fn still takes iRef by blank intent, but iRef's
type is a _ref:

  proc begin_fn(<blank> iRef: _ref(int(64))) {
    iRef = 5;
    writeln(iRef);
  }

Therefore it is the pointer (that is stored in the outer iRef),
not the pointee, that is passed into begin_fn. Its being a ref
fools the compiler into allowing iRef=5. At run time, that updates
the pointee of the outer iRef.

THE FIX is to check for reference types in handleCaptureArgs (see
r21940). When the type of a variable that is being passed into the
task function is a _ref of a type that is subject to capture (e.g.
ints, records, etc.):

- Replace the type of the task function's corresponding formal
with the value type.

- Replace the actual with a temp that holds a dereference of
that actual.

- Update the corresponding CallInfo appropriately.

>> functionResolution.cpp
 handleCaptureArgs
 replaceActualWithDeref
 resolveNormalCall

I ignored the cases when the dereference type is not subject to
capture, e.g. it is an array or a synchronization type. That's
because I have not seen any issues in those cases and prefered
to leave the behavior unchanged for simplicity.

>> type.cpp
  needsCapture(Type*)

Discussion: It is almost enough to just replace the type of the task
function's formal (i.e. the first of the above actions). However,
this results in a coercion wrapper around the call to the task function.

As a result, handleCaptureArgs() is invoked once again on that task
function, although now with a different CallExpr. This is OK, especially
there is still only a single call site per task function, i.e. it
is still invoked only once.

But unfortunately this interferes with lowerIterators. I observed
PRIM_YIELD calls surviving until codegen, e.g., in release/ssca2.
I did not investigate, but I had flashbacks to my work on r21426.
There, a major challenge was go get traversals of an iterator examine
the now-outlined task functions - iterator yields could now
end up in those functions, but they still needed to be found
and handled properly.

Replacing the actual with a dereference seemed easier than
investigating the yield issue, so that's what I did.

As a side benefit, in some cases references to variables
used to be captured (resulting in indirections within
the task functions, I expect), and now values are
captured instead (correct and more efficient).

BTW TODO I suggest making a similar change in coercionWrap().
That is, in the case of ref-deref difference, replace the actual
using replaceActualWithDeref(), rather than handle it in the wrapper.
This will obviate some wrappers altogether.

THE CLARIFICATION is that an index variable of a coforall loop
is not subject to capture for the loop body task.
It is still subject to capture (depending on its type)
for any task construct *within* that loop body, e.g. a
nested coforall.

To implement that, I extended findOuterVars to drop such variables.
I added isCorrespCoforallIndex() to detect whether a variable is
an index of a coforall loop, relying on FLAG_COFORALL_INDEX_VAR
and matching the parentExpr of the variable's DefExpr vs.
that of the task construct's block/task function.

A more robust check would also verify that the index variable
comes from the _indexOfInterest. Given that this is a bit of
work in the case where the coforall is zippered, I chose to
conclude that that was not necessary.

>> createTaskFunctions.cpp
 isCorrespCoforallIndex
 findOuterVars

TEST CHANGES were also necessary because in a couple of places
the assigning-to-constant (aka "illegal lvalue in assignment")
check used to being fooled by the captured variable's ref-ness.
ref-intent-bug-4begin.chpl now works under all configurations,
so I removed the .skipif.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22008 3a8e244f-b0f2-452b-bcba-4c88e055c3ca