Use param localeModelHasSublocales in range and DefaultRectangular leaders
[Reviewed by Sung]

Instead of using (numSublocs != 0) to decide if there are sublocales, use
(localeModelHasSublocales && numSublocs != 0).  This allows the choice to be
made at compile time when localeModelHasSublocales is false, as is the case in
the flat locale model. This avoids a small memory leak regression in
memory/figueroa/LeakedMemory5 and also reduces code size for parallel iterators
in the flat model.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22014 3a8e244f-b0f2-452b-bcba-4c88e055c3ca