Move the removeRandomJunk() pass to after insertRuntimeTypeTemps() in resolution
[Reviewed by Sung, Tom]

The pass insertRuntimeTypeTemps() can insert calls to methods that include
methodTokens. One of the things removeRandomJunk() removes is methodTokens and
we expect them to not exist anymore after it is run. Reorder these two
passes so the methodTokens get cleaned up. This should fix codes that use the
HDFS module.

I also added test/arrays/diten/distArrInRecord.chpl, which declares a
distributed array as a record field and fails in the same way as programs that
use HDFS prior to this revision.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22016 3a8e244f-b0f2-452b-bcba-4c88e055c3ca