Fixed a bug in heap allocation (exhibited by parallel/taskPar/figueroa/taskParallel.chpl under --local).
[reviewed by Sung]

The bug arises (under --local && !NUMA) when a global variable is passed
as an actual to a task function, and the corresponding task function formal
has a _ref type.

Given that:
- variables listed in a 'ref' clause do not count,
- arrays/domains/distribs are excluded,
- syncs/singles are not passed with _ref (the formal has a _syncvar type),

this was observed only for atomics, since r21924.

The bug is that makeHeapAllocations() was assuming that such a global
was heap-allocated, and so changing accesses to the actual value to
accesses via the global's _value field.

MHA delegates the actual heap allocation to insertWideReferences().
There, it happens only if requireWideReferences(),
which is false under --local && !NUMA.

As a result, accessing via the _value field fails the check that
the global itself is not NULL.

Since under --local && !NUMA there is no need to heap-allocate
globals, I made makeHeapAllocations() not modify accesses to
use _value for globals.

TODO: if globals do not need heap allocation when !needHeapVars()
(i.e., currently, when --local or UGNI or GASNET-everything -
a wider set of configurations than requireWideReferences()),
we should change the logic in insertWideReferences() w.r.t.
heapAllocateGlobals/chpl__heapAllocateGlobals().

The actual code change in this commit is 5 lines.
The rest are comments, including the following,
which is my best-effort summary of the current state:

Notes on
  makeHeapAllocations()    //invoked from parallel()
  insertWideReferences()

------
Terminology/abbreviations:

'on'+'begin' is considered to be both an 'on' and a 'begin'
- see also FLAG_NON_BLOCKING

A "global" is a module-level symbol, usu. a VarSymbol.
A 'var' is a global if, equivalently:
 isModuleSymbol(var->defPoint->parentSymbol)
 isGlobal(var)

MHA = makeHeapAllocations() and functions it invokes
IWR = insertWideReferences() and functions it invokes
------

MHA and IWR take care of the following, among others:
- heap allocation for remote access
- heap allocation for 'begin'
- change acces to variable -> access to its ._value
- set up wide references
- broadcasting of globals

In more details:

Heap allocation for remote access is done:
- for globals - in IWR, if:
   requireWideReferences()
- for a local - in MHA, if:
   needHeapVars() && the local can be passed to an 'on'

Heap allocation for 'begin' is done:
- for globals - n/a
   see above instead
- for a local - in MHA, if:
   the local can be passed to a 'begin'

Change acces to variable -> access to its ._value
- for globals - in MHA, if:
   requireWideReferences()
- for locals - in MHA, if:
   the local is subject to heap allocation
   in either of the above two categories

Wide references are set up in IWR, if:
   requireWideReferences()

Broadcasting:
- of certain global constants       - in MHA/findHeapVarsAndRefs, if:
   !fLocal
- of global arrays/domains/distribs - in MHA/findHeapVarsAndRefs, if:
   !fLocal
- of locations of the other globals - in IWR, if:
   requireWideReferences()


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22018 3a8e244f-b0f2-452b-bcba-4c88e055c3ca