Added small-int operators to the language spec
[reviewed by Elliot]

This commit adds the new operators on small ints (int(8), uint(8), int(16),
uint(16)).  In doing so, I typically continued the current verbose approach
of listing out signatures rather than using the generic form that we
actually use in the implementation.  Primarily this was because it was a
smaller change by preserving the original spirit; in part it's because I
think it's almost easier to process as a reader in this form, in spite of
the verbosity.

In the case of logical operators, I switched to the generic form, however,
because it went from a linear enumeration of integer types at each width to
a cross-product of int/uint combinations.

While here, I removed int/uint pairs that (a) I don't believe we support and
(b) didn't specify their return types anyway.

I also fixed the linefeeds in a printchploutput environment while here.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22067 3a8e244f-b0f2-452b-bcba-4c88e055c3ca